implement login for admin, students, responsible

+++ src/controllers/administrador.controller.ts
@@ -2,7 +2,7 @@ import { Request, Response } from "express";
 import { AdministradorService } from "../services/administrador.service";
 import { HttpResponse } from "../shared/response/http.response";
 import { DeleteResult, UpdateResult } from "typeorm";
-import { AuthAdminService } from "../auth/services/authAdmin.service";
+import { AuthAdminService } from "../auth/services/auth.service";
 
 export class AdministradorController{
     constructor(
@@ -13,16 +13,19 @@ export class AdministradorController{
 
     async getAdministradors(req: Request, res: Response) {
         try {
+            // Valid token
             const token = req.cookies.accessToken
             if (!token) {
                 return this.httpResponse.Unauthorized(res,"No tienes permisos");
             }
 
+            // Valid super
             const validSuper = await this.auth.verifiedIsSuper(token);
 
             if (!validSuper) {
                 return this.httpResponse.Unauthorized(res,"No tiene permisos")
             }
+
 
             const data = await this.administradorService.findAll();
             if (data.length === 0) {
@@ -36,8 +39,21 @@ export class AdministradorController{
     }
 
     async getAdministradorById(req: Request, res: Response) {
+        
         const {id}= req.params;
         try {
+            // Valid token
+            const token = req.cookies.accessToken
+            if (!token) {
+                return this.httpResponse.Unauthorized(res,"No tienes permisos");
+            }
+    
+            // Valid super
+            const validSuper = await this.auth.verifiedIsSuper(token);
+    
+            if (!validSuper) {
+                return this.httpResponse.Unauthorized(res,"No tiene permisos")
+            }
             const data = await this.administradorService.findbyid(id);
             if (!data) {
                 return this.httpResponse.NotFound(res, "No existe datos")
@@ -49,7 +65,20 @@ export class AdministradorController{
     }
 
     async createAdministrador(req: Request, res: Response) {
+        
         try {
+            // Valid token
+            const token = req.cookies.accessToken
+            if (!token) {
+                return this.httpResponse.Unauthorized(res,"No tienes permisos");
+            }
+    
+            // Valid super
+            const validSuper = await this.auth.verifiedIsSuper(token);
+    
+            if (!validSuper) {
+                return this.httpResponse.Unauthorized(res,"No tiene permisos")
+            }
             const data = await this.administradorService.create(req.body);
             if (data) {
                 return this.httpResponse.Ok(res, data)
@@ -62,8 +91,21 @@ export class AdministradorController{
     }
 
     async updateAdministrador(req: Request, res: Response) {
+        
         const {id}= req.params;
         try {
+            // Valid token
+            const token = req.cookies.accessToken
+            if (!token) {
+                return this.httpResponse.Unauthorized(res,"No tienes permisos");
+            }
+    
+            // Valid super
+            const validSuper = await this.auth.verifiedIsSuper(token);
+    
+            if (!validSuper) {
+                return this.httpResponse.Unauthorized(res,"No tiene permisos")
+            }
             const data: UpdateResult = await this.administradorService.update(id,req.body);
 
             if (!data.affected) {
@@ -77,8 +119,21 @@ export class AdministradorController{
     }
 
     async deteleAdministrador(req: Request, res: Response) {
+        
         const {id}= req.params;
         try {
+            // Valid token
+            const token = req.cookies.accessToken
+            if (!token) {
+                return this.httpResponse.Unauthorized(res,"No tienes permisos");
+            }
+    
+            // Valid super
+            const validSuper = await this.auth.verifiedIsSuper(token);
+    
+            if (!validSuper) {
+                return this.httpResponse.Unauthorized(res,"No tiene permisos")
+            }
             const data: DeleteResult = await this.administradorService.delete(id);
 
             if (!data.affected) {
@@ -91,10 +146,11 @@ export class AdministradorController{
         }
     }
 
-
     async findAdministradorWithTipo(req: Request, res: Response) {
         const {id}= req.params;
         try {
+            
+
             const data = await this.administradorService.findAdminWithTipo(id);
             if (!data) {
                 return this.httpResponse.NotFound(res, "No existe datos")
@@ -104,4 +160,6 @@ export class AdministradorController{
             return this.httpResponse.NotFound(res, e)
         }
     }
+
+    
 }
